% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gale_Shapley.R
\name{Gale_Shapley}
\alias{Gale_Shapley}
\title{This is the version of Gale-Shapley stable matching algorithm (translated from the Matlab code in Menzel (2015)).}
\usage{
Gale_Shapley(U, V, return.data.frame = FALSE, cpp = TRUE)
}
\arguments{
\item{U}{The utility matrix for the women's side. Each row is a woman, each column is a man.
The matrix entry (i,j) is the utility that woman \code{i} gains from pairing with man \code{j}. 
In other words, the utility is computed from woman \code{i}'s perspective.}

\item{V}{The utility matrix for the men's side. Each column is a man, each row is a woman.
The matrix entry (i,j) is the utility that man \code{j} gains from pairing with woman \code{i}. 
In other words, the utility is computed from man \code{j}'s perspective.}

\item{return.data.frame}{logical Should a \code{data.frame} of the matching be returned instead of the
paring matrix mu?}

\item{cpp}{logical Should the \code{Rcpp} version of the code be used. This is much faster and uses a lot less memory.}
}
\value{
The function return depends on the \code{return.data.frame} value. 
If TRUE, it returns
\item{data.frame}{a two-column \code{data.frame} with the first column a women's index and the second column the 
men's index of their partner. It has as many rows as there are partnerships.}
If FALSE, it returns the following matrix: 
\item{mu}{If \code{cpp=TRUE}, a vector of length the number of women (\code{nrow(U)}) with the 
index of the matching man (i.e., the index is the row in \code{V} of the man). If there is no
matching man, the index is 0. This can be used to reconstruct the matching matrix.
If \code{cpp=FALSE}, the matching matrix, where 1 represents a pairing, 0 otherwise. 
Each row is a woman, each column is a man. The order of the rows is the same as the 
rows in \code{U}. The order of the columns is the same as the columns in \code{V}.}
}
\description{
This code allows the self-matched option
}
\references{
Menzel, Konrad (2015).
\emph{Large Matching Markets as Two-Sided Demand Systems}
Econometrica, Vol. 83, No. 3 (May, 2015), 897-941.
}
\seealso{
rpm
}
\keyword{models}
