% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.utilities.R
\name{list_rhs.formula}
\alias{list_rhs.formula}
\title{Returns a list containing the terms in a given formula}
\usage{
list_rhs.formula(object)
}
\arguments{
\item{object}{formula A formula having a right-hand-side that can be interpretated as a rpm specification.
returns a list containing terms in a given
formula, handling \code{+} and \code{-} operators and parentheses, and
keeping track of whether a term has a plus or a minus sign.}
}
\value{
\code{list_rhs.formula} returns a list of formula terms, with an additional numerical vector attribute \code{"sign"} with of the same length, giving the corresponding term's sign as \code{+1} or \code{-1}.
}
\description{
Returns a list containing the terms in a given formula
}
