% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.rpm.R
\name{logLikNull}
\alias{logLikNull}
\alias{logLikNull.rpm}
\title{Calculate the null model likelihood}
\usage{
logLikNull(object, ...)

\method{logLikNull}{rpm}(object, ...)
}
\arguments{
\item{object}{a fitted model.}

\item{...}{further arguments to lower-level functions.

\code{logLikNull} computes, when possible the log-probability of
the data under the null model (reference distribution).}
}
\value{
\code{logLikNull} returns an object of type \code{\link{logLik}} if it is
able to compute the null model probability, and \code{NA} otherwise.
}
\description{
Calculate the null model likelihood
}
\section{Methods (by class)}{
\itemize{
\item \code{logLikNull(rpm)}: A method for [`rpm`] fits to compute the null likelihood (that is, relative to the constant only model).

}}
