% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpm.utilities.R
\name{message_print}
\alias{message_print}
\title{[`print`] objects to the [`message`] output.}
\usage{
message_print(..., messageArgs = NULL)
}
\arguments{
\item{...}{arguments to [`print`].}

\item{messageArgs}{a list of arguments to be passed directly to [`message`].}
}
\description{
A thin wrapper around [`print`] that captures its output and prints
it as a [`message`], usually to STDERR.
Tis is part of [`statnet.common`].
}
\examples{
cat(1:5)

print(1:5)
message_print(1:5) # Looks the same (though may be in a different color on some frontends).

suppressMessages(print(1:5)) # Still prints
suppressMessages(message_print(1:5)) # Silenced
}
