% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpm-internal.R
\name{rpm-internal}
\alias{rpm-internal}
\alias{GSi_TU}
\alias{GSi_NTU}
\alias{GS_TU}
\alias{GS_NTU}
\alias{PSeqcond}
\alias{PSgeqcond}
\alias{augpmf}
\alias{augpmfWM}
\alias{augpmfnew}
\alias{checkstable}
\alias{eqcond}
\alias{geqcond}
\alias{gloglik}
\alias{hloglik}
\alias{loglik}
\alias{probcounts}
\alias{qloglik}
\alias{stablearma}
\title{Internal rpm Objects}
\description{
Internal \code{\link[=rpm-package]{rpm}} functions.
}
\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written.
}
\seealso{
rpm-package
}
\keyword{internal}
