% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpm.model.functions.R
\name{rpm.model.functions}
\alias{rpm.model.functions}
\title{Creates a model function list for the continuous terms in a Revealed Preference Matchings Model}
\usage{
rpm.model.functions(model.terms, control)
}
\arguments{
\item{model.terms}{For the details on the possible countinuous \code{<model terms>}, see
\code{\link{rpm-terms}}. This includes the 
covariates used to construct the model matrix.
They are used in conjunction with the model terms.}

\item{control}{A list of control parameters for algorithm tuning. Constructed using
\code{\link{control.rpm}}, which should be consulted for specifics.}
}
\value{
A list of model terms as bivariate functions.
}
\description{
\code{\link{rpm.model.matrix}} assumes a bipartite network (i.e. two-sided matching market)
It creates a model matrix according to the formula passed in.
See \code{\link{rpm-terms}} for a description of the possible terms.
}
\examples{
# nothing yet
}
\references{
Menzel, Konrad (2015).
\emph{Large Matching Markets as Two-Sided Demand Systems}
Econometrica, Vol. 83, No. 3 (May, 2015), 897-941.
}
\seealso{
rpm
}
\keyword{models}
